requirements.txt
# QuantSwarm v4 — Python Dependencies
# Install: pip install -r requirements.txt
# Python >= 3.11 required

# Core ML
torch>=2.2.0
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0

# NLP
transformers>=4.38.0
tokenizers>=0.15.0
sentence-transformers>=2.5.0

# Data sources
praw>=7.7.0                    # Reddit API (free)
sec-edgar-downloader>=5.0.0    # SEC filings
feedparser>=6.0.0              # RSS news
yfinance>=0.2.37               # Market data (free)
ccxt>=4.2.0                    # Crypto exchange data
httpx>=0.27.0                  # Async HTTP (Etherscan, etc.)

# Explainability
shap>=0.44.0

# FastAPI dashboard
fastapi>=0.110.0
uvicorn>=0.27.0
websockets>=12.0
jinja2>=3.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Backtesting
# (walk_forward uses yfinance + pandas directly — no extra dep needed)

# Alerts
python-telegram-bot>=21.0.0
requests>=2.31.0

# Config
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
loguru>=0.7.0
tenacity>=8.2.0                # retry logic
schedule>=1.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
hypothesis>=6.98.0

# Type hints
pydantic>=2.6.0
